+++ src/vkgs/engine/engine.cc
@@ -26,20 +26,20 @@
 #include <vkgs/util/clock.h>
 
 #include "vkgs/engine/splat_load_thread.h"
-#include "vkgs/engine/vulkan/context.h"
-#include "vkgs/engine/vulkan/swapchain.h"
-#include "vkgs/engine/vulkan/attachment.h"
-#include "vkgs/engine/vulkan/descriptor_layout.h"
-#include "vkgs/engine/vulkan/pipeline_layout.h"
-#include "vkgs/engine/vulkan/compute_pipeline.h"
-#include "vkgs/engine/vulkan/graphics_pipeline.h"
-#include "vkgs/engine/vulkan/render_pass.h"
-#include "vkgs/engine/vulkan/framebuffer.h"
-#include "vkgs/engine/vulkan/descriptor.h"
-#include "vkgs/engine/vulkan/buffer.h"
-#include "vkgs/engine/vulkan/cpu_buffer.h"
-#include "vkgs/engine/vulkan/uniform_buffer.h"
-#include "vkgs/engine/vulkan/shader/uniforms.h"
+#include "vkgs/vulkan/context.h"
+#include "vkgs/vulkan/swapchain.h"
+#include "vkgs/vulkan/attachment.h"
+#include "vkgs/vulkan/descriptor_layout.h"
+#include "vkgs/vulkan/pipeline_layout.h"
+#include "vkgs/vulkan/compute_pipeline.h"
+#include "vkgs/vulkan/graphics_pipeline.h"
+#include "vkgs/vulkan/render_pass.h"
+#include "vkgs/vulkan/framebuffer.h"
+#include "vkgs/vulkan/descriptor.h"
+#include "vkgs/vulkan/buffer.h"
+#include "vkgs/vulkan/cpu_buffer.h"
+#include "vkgs/vulkan/uniform_buffer.h"
+#include "vkgs/vulkan/shader/uniforms.h"
 
 #include "generated/parse_ply_comp.h"
 #include "generated/projection_comp.h"
