+++ src/vkgs/vulkan/buffer.h
@@ -1,12 +1,12 @@
-#ifndef VKGS_ENGINE_VULKAN_BUFFER_H
-#define VKGS_ENGINE_VULKAN_BUFFER_H
+#ifndef VKGS_VULKAN_BUFFER_H
+#define VKGS_VULKAN_BUFFER_H
 
 #include <memory>
 #include <vector>
 
 #include <vulkan/vulkan.h>
 
-#include "vkgs/engine/vulkan/context.h"
+#include "vkgs/vulkan/context.h"
 
 namespace vkgs {
 namespace vk {
@@ -41,4 +41,4 @@ class Buffer {
 }  // namespace vk
 }  // namespace vkgs
 
-#endif  // VKGS_ENGINE_VULKAN_BUFFER_H
+#endif  // VKGS_VULKAN_BUFFER_H
