+++ src/vkgs/vulkan/cpu_buffer.h
@@ -1,11 +1,11 @@
-#ifndef VKGS_ENGINE_VULKAN_CPU_BUFFER_H
-#define VKGS_ENGINE_VULKAN_CPU_BUFFER_H
+#ifndef VKGS_VULKAN_CPU_BUFFER_H
+#define VKGS_VULKAN_CPU_BUFFER_H
 
 #include <memory>
 
 #include <vulkan/vulkan.h>
 
-#include "vkgs/engine/vulkan/context.h"
+#include "vkgs/vulkan/context.h"
 
 namespace vkgs {
 namespace vk {
@@ -31,4 +31,4 @@ class CpuBuffer {
 }  // namespace vk
 }  // namespace vkgs
 
-#endif  // VKGS_ENGINE_VULKAN_CPU_BUFFER_H
+#endif  // VKGS_VULKAN_CPU_BUFFER_H
