+++ src/vkgs/vulkan/descriptor.h
@@ -1,12 +1,12 @@
-#ifndef VKGS_ENGINE_VULKAN_DESCRIPTOR_H
-#define VKGS_ENGINE_VULKAN_DESCRIPTOR_H
+#ifndef VKGS_VULKAN_DESCRIPTOR_H
+#define VKGS_VULKAN_DESCRIPTOR_H
 
 #include <memory>
 
 #include <vulkan/vulkan.h>
 
-#include "vkgs/engine/vulkan/context.h"
-#include "vkgs/engine/vulkan/descriptor_layout.h"
+#include "vkgs/vulkan/context.h"
+#include "vkgs/vulkan/descriptor_layout.h"
 
 namespace vkgs {
 namespace vk {
@@ -33,4 +33,4 @@ class Descriptor {
 }  // namespace vk
 }  // namespace vkgs
 
-#endif  // VKGS_ENGINE_VULKAN_DESCRIPTOR_H
+#endif  // VKGS_VULKAN_DESCRIPTOR_H
