+++ src/vkgs/vulkan/descriptor_layout.cc
@@ -1,4 +1,4 @@
-#include "vkgs/engine/vulkan/descriptor_layout.h"
+#include "vkgs/vulkan/descriptor_layout.h"
 
 #include <unordered_map>
