+++ src/vkgs/vulkan/graphics_pipeline.h
@@ -1,14 +1,14 @@
-#ifndef VKGS_ENGINE_VULKAN_GRAPHICS_PIPELINE_H
-#define VKGS_ENGINE_VULKAN_GRAPHICS_PIPELINE_H
+#ifndef VKGS_VULKAN_GRAPHICS_PIPELINE_H
+#define VKGS_VULKAN_GRAPHICS_PIPELINE_H
 
 #include <memory>
 #include <string>
 
 #include <vulkan/vulkan.h>
 
-#include "vkgs/engine/vulkan/context.h"
-#include "vkgs/engine/vulkan/pipeline_layout.h"
-#include "vkgs/engine/vulkan/shader_module.h"
+#include "vkgs/vulkan/context.h"
+#include "vkgs/vulkan/pipeline_layout.h"
+#include "vkgs/vulkan/shader_module.h"
 
 namespace vkgs {
 namespace vk {
@@ -48,4 +48,4 @@ class GraphicsPipeline {
 }  // namespace vk
 }  // namespace vkgs
 
-#endif  // VKGS_ENGINE_VULKAN_GRAPHICS_PIPELINE_H
+#endif  // VKGS_VULKAN_GRAPHICS_PIPELINE_H
